+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/utils/utils.py
@@ -9,6 +9,9 @@ from utils.column_values.categorical_column_values import (
     MainJobRoleLabels,
     PrimaryServiceType,
 )
+from utils.column_values.categorical_columns_by_dataset import (
+    EstimatedIndCQCFilledPostsByJobRoleCategoricalValues as CatVals,
+)
 from utils.value_labels.ascwds_worker.ascwds_worker_jobgroup_dictionary import (
     AscwdsWorkerValueLabelsJobGroup,
 )
@@ -67,6 +70,19 @@ def add_job_role_groups_column(
 
 
 class HistoricJobRoleAdjustmentConfig:
+    """
+    A dictionary in which keys are dates before which we reallocate job roles.
+        Date: {
+            historic role to reallocate: {
+                receiving role: amount to receive,
+            },
+        }
+
+    Please note the registered manager role must not be a receiving role due to
+    the order of functions in _04_estimate job roles script (historical reallocation
+    happens after RM reassignment). This role is taken from CQC counts and therefore
+    must be that count thereafter.
+    """
 
     adjustment_dict = {
         date(2023, 8, 1): {
@@ -126,8 +142,12 @@ class CategoricalColumnTypes:
         pl.Categories("provider", namespace="filled_posts")
     )
     BrandCatType = pl.Categorical(pl.Categories("brand", namespace="filled_posts"))
-    JobRoleEnumType = pl.Enum(AscwdsWorkerValueLabelsJobGroup.all_roles())
-    JobGroupEnumType = pl.Enum(AscwdsWorkerValueLabelsJobGroup.all_job_groups())
+    JobRoleEnumType = pl.Enum(
+        CatVals.main_job_role_labels_column_values.categorical_values
+    )
+    JobGroupEnumType = pl.Enum(
+        CatVals.main_job_group_labels_column_values.categorical_values
+    )
     EstimatesFilledPostSourceEnumType = pl.Enum(
         [
             EstimateFilledPostsSource.imputed_pir_filled_posts_model,