+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -49,8 +49,8 @@ from utils.column_values.categorical_column_values import (
     JobGroupLabels,
     JobRoleFilteringRule,
 )
-from utils.value_labels.ascwds_worker.ascwds_worker_jobgroup_dictionary import (
-    AscwdsWorkerValueLabelsJobGroup,
+from utils.column_values.categorical_columns_by_dataset import (
+    EstimatedIndCQCFilledPostsByJobRoleCategoricalValues as CatVals,
 )
 
 
@@ -1474,9 +1474,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
         {
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
             IndCQC.registered_manager_count: pl.Float32,
         }
     )
@@ -1484,9 +1484,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
         {
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
             IndCQC.estimate_filled_posts_by_job_role_manager_adjusted: pl.Float32,
         }
     )
@@ -1495,9 +1495,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
         {
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
             IndCQC.registered_manager_count: pl.Float32,
             IndCQC.difference_between_estimate_and_cqc_registered_managers: pl.Float32,
         }
@@ -1507,9 +1507,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
         {
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
             IndCQC.proportion_of_non_rm_managerial_estimated_filled_posts_by_role: pl.Float32,
         }
     )
@@ -1518,9 +1518,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
         {
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
             IndCQC.registered_manager_count: pl.Int32,
             IndCQC.difference_between_estimate_and_cqc_registered_managers: pl.Float32,
             IndCQC.proportion_of_non_rm_managerial_estimated_filled_posts_by_role: pl.Float32,
@@ -1533,9 +1533,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
             IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.estimate_filled_posts: pl.Float32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role_manager_adjusted: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
             IndCQC.estimate_filled_posts_from_all_job_roles: pl.Float32,
             IndCQC.difference_estimate_filled_posts_and_from_all_job_roles: pl.Float32,
         }
@@ -1547,9 +1547,9 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
             IndCQC.location_id: pl.String,
             IndCQC.cqc_location_import_date: pl.Date,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
-                AscwdsWorkerValueLabelsJobGroup.all_roles()
+                CatVals.main_job_role_labels_column_values.categorical_values
             ),
-            IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
+            IndCQC.estimate_filled_posts_by_job_role_manager_adjusted: pl.Float32,
             IndCQC.estimate_filled_posts_by_job_role_historically_reallocated: pl.Float32,
         }
     )
@@ -1707,9 +1707,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }
@@ -1718,9 +1716,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }
@@ -1729,9 +1725,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }
@@ -1740,9 +1734,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }
@@ -1751,9 +1743,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }
@@ -1762,9 +1752,7 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
         IndCQC.main_job_role_clean_labelled: CatColType.JobRoleEnumType,
-        IndCQC.main_job_group_labelled: pl.Enum(
-            AscwdsWorkerValueLabelsJobGroup.all_job_groups()
-        ),
+        IndCQC.main_job_group_labelled: CatColType.JobGroupEnumType,
         IndCQC.ascwds_job_role_counts: pl.Int64,
         IndCQC.job_role_filtering_rule: CatColType.JobRoleFilteringRuleCatType,
     }