+++ CHANGELOG.md
@@ -17,6 +17,10 @@ All notable changes to this project will be documented in this file.
 - Added folder structure for workforce characteristics that includes starters, leavers and vacancies merge job.
   Also added merge job and it's validation to SLV step function, and its required terraform changes.
 
+- Added clean job, its validation and test scripts within SLV pipeline. Updated the required terraform changes for the same.
+
+- Added impute job, its validation and test scripts within SLV pipeline. Updated the required terraform changes for the same.
+
 ### Changed
 - Refactored job group filter to use magic numbers for outlier bounds.
 
@@ -24,7 +28,10 @@ All notable changes to this project will be documented in this file.
 
 - Refactored job group filter to work for brands and providers as well as locations.
 
+- Changed _04_estimate job role filled posts so historical reallocation is called after manager adjustment.
+
 ### Fixed
+- Replace the all roles and all job groups functions with the appropriate categorical values attribute.
 
 
 ## [v2026.05.0] - 12/06/2026